requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
opencv-python
psycopg2-binary
pytest
awsebcli
python-multipart
numpy
aiofiles
httpx